requirements.txt
openai-whisper==20231117
torch
torchaudio
ffmpeg-python
numpy
Pillow
pyinstaller
tiktoken
